requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pydantic
